move method stringToValue to super class and keep only the special cases in BCStyle and RFC4519Style
